+++ simplepyml/util/functions/activation/relu.py
@@ -1,4 +1,4 @@
-from .activation import BaseActivation
+from simplepyml.util.functions.activation.activation import BaseActivation
 import numpy as np
 
 class Relu(BaseActivation):
@@ -9,7 +9,7 @@ class Relu(BaseActivation):
         self,
         x: int | float | np.integer | np.floating | list | np.ndarray
     ) -> np.float64 | np.ndarray:
-        y = np.asarray(x)
+        y = np.asarray(x, dtype=np.float64)
         y[y < 0] = 0
         return y
     
@@ -17,7 +17,7 @@ class Relu(BaseActivation):
         self,
         x: int | float | np.integer | np.floating | list | np.ndarray
     ) -> np.float64 | np.ndarray:
-        y = np.asarray(x)
+        y = np.asarray(x, dtype=np.float64)
         y[y <= 0] = 0
         y[y > 0] = 1
         return y
